LayoutTests/fast/selectors/not-backtracking-expected.txt

Check the basic features of the ":not(selectorList)" pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "testcase:not(bb > aa > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > aa > bb testcase)').length is 0
PASS styledElements is []

Testing "testcase:not(bb > aa > aa > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > aa > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > aa > bb > aa testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > aa > bb > aa testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:not(bb > * > aa > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:not(bb > * > aa > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(bb > * > aa > bb > aa testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(bb > * > aa > bb > aa testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:not(bb > aa > * > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > * > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > * > bb > aa testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(bb > aa > * > bb > aa testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:not(bb > * > * > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:not(bb > * > * > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(bb > * > * > bb > aa testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(bb > * > * > bb > aa testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:not(* > * > * > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* > * > * > bb > aa testcase)').length is 0
PASS styledElements is []

Testing "testcase:not(aa > aa > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > aa > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > aa > bb testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > aa > bb testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:not(aa > aa > * > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > * > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > * > bb testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > * > bb testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:not(aa > * > * > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:not(aa > * > * > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(aa > * > * > bb testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(aa > * > * > bb testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:not(aa > aa > aa > * testcase)"
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > aa > * testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > aa > * testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(aa > aa > aa > * testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:not(aa > * > * > * testcase)"
PASS document.querySelectorAll('#test-root testcase:not(aa > * > * > * testcase)').length is 0
PASS styledElements is []

Testing "testcase:not(* > * > * > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* > * > * > bb testcase)').length is 0
PASS styledElements is []

Testing "testcase:not(* > * > * > * testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* > * > * > * testcase)').length is 0
PASS styledElements is []

Testing "testcase:not(dd + dd + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + dd ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + dd ~ testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + dd ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase4"]

Testing "testcase:not(dd + dd + cc + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + * ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + * ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + * ~ testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + cc + * ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase4"]

Testing "testcase:not(dd + dd + * + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + dd ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + dd ~ testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + dd ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase4"]

Testing "testcase:not(dd + dd + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + * ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + * ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + * ~ testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(dd + dd + * + * ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase4"]

Testing "testcase:not(dd + * + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(dd + * + cc + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(dd + * + cc + dd ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(dd + * + cc + dd ~ testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "testcase:not(* + dd + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* + dd + cc + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(* + dd + cc + dd ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(* + dd + cc + dd ~ testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "testcase:not(* + * + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* + * + cc + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(* + * + cc + dd ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(* + * + cc + dd ~ testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "testcase:not(cc + cc + cc + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + dd ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + dd ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + dd ~ testcase)')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + dd ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase2","testcase3","testcase4"]

Testing "testcase:not(cc + cc + cc + cc + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + * ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + * ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + * ~ testcase)')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + cc + * ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase2","testcase3","testcase4"]

Testing "testcase:not(cc + cc + cc + * + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + dd ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + dd ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + dd ~ testcase)')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + dd ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase2","testcase3","testcase4"]

Testing "testcase:not(cc + cc + cc + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + * ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + * ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + * ~ testcase)')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + cc + * + * ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase2","testcase3","testcase4"]

Testing "testcase:not(cc + cc + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + * + * + * ~ testcase)').length is 3
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + * + * + * ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + * + * + * ~ testcase)')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:not(cc + cc + * + * + * ~ testcase)')[2].id is "testcase4"
PASS styledElements is ["testcase2","testcase3","testcase4"]

Testing "testcase:not(cc + * + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(cc + * + * + * + * ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(cc + * + * + * + * ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(cc + * + * + * + * ~ testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "testcase:not(* + cc + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* + cc + * + * + * ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(* + cc + * + * + * ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(* + cc + * + * + * ~ testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "testcase:not(* + * + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:not(* + * + * + * + * ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:not(* + * + * + * + * ~ testcase)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:not(* + * + * + * + * ~ testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(ee) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(ee, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(ee) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(ee) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(ee, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(ee) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(ee) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(ee, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(ee) > * > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee) > * > ee > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :not(ee) > * > ee > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee) > * > ee > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee, foo, bar, baz) > * > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > ee > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > ee > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > ee > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee) > * > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :not(ee) > * > * > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee, foo, bar, baz) > * > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee) > * > * > ee > * testcase"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > ee > * testcase').length is 2
PASS document.querySelectorAll('#test-root :not(ee) > * > * > ee > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > ee > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee, foo, bar, baz) > * > * > ee > * testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > ee > * testcase').length is 2
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > ee > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > ee > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":not(ee) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > * > * testcase').length is 4
PASS document.querySelectorAll('#test-root :not(ee) > * > * > * > * testcase')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > * > * testcase')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > * > * testcase')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(ee) > * > * > * > * testcase')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing ":not(ee, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > * > * testcase').length is 4
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > * > * testcase')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(ee, foo, bar, baz) > * > * > * > * testcase')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing ":not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :not(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase').length is 4
PASS document.querySelectorAll('#test-root :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing ":not(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :not(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase').length is 4
PASS document.querySelectorAll('#test-root :not(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :not(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root :not(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root :not(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "bb > :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root bb > :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase').length is 3
PASS document.querySelectorAll('#test-root bb > :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root bb > :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root bb > :not(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[2].id is "testcase4"
PASS styledElements is ["testcase2","testcase3","testcase4"]

PASS successfullyParsed is true

TEST COMPLETE
